LayoutTests/http/tests/navigation/post-frames.html

<script>
  if (window.layoutTestController) {
      layoutTestController.dumpAsText();
      layoutTestController.dumpChildFramesAsText();
      layoutTestController.waitUntilDone();
      layoutTestController.clearBackForwardList();
      layoutTestController.dumpBackForwardList();
   }
</script>  

<p>Tests that a POST request in a frame is handled correctly. If this test passes, you will see a page indicating a form was POSTed.</p>

<iframe src="resources/page-that-posts.html"></iframe>
