Safari-611.1.10/LayoutTests/http/tests/navigation/post-frames-expected.txt

Tests that a POST request in a frame is handled correctly. If this test passes, you will see a page indicating a form was POSTed.



--------
Frame: '<!--frame1-->'
--------
This page was requested with the HTTP method POST.

Parameters:

the-input = input value goes here
submitwithpost = Submit with POST

============== Back Forward List ==============
        http://127.0.0.1:8000/navigation/post-frames.html  **nav target**
            http://127.0.0.1:8000/navigation/resources/page-that-posts.html (in frame "<!--frame1-->")
curr->  http://127.0.0.1:8000/navigation/post-frames.html
            http://127.0.0.1:8000/navigation/resources/form-target.pl (in frame "<!--frame1-->")  **nav target**
===============================================
